+++ odoo/local-src/pricelist_discount/models/purchase_order.py
@@ -2,7 +2,7 @@
 # Copyright 2017 Julien Coux (Camptocamp)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo import models, fields, api
+from odoo import models, fields
 
 
 class PurchaseOrder(models.Model):
@@ -10,15 +10,6 @@ class PurchaseOrder(models.Model):
 
     supplier_promotion_allowed = fields.Boolean(
         string='Supplier promotion allowed',
-        states={
-            'purchase': [('readonly', True)],
-            'done': [('readonly', True)],
-            'cancel': [('readonly', True)],
-        },
+        related='partner_id.supplier_promotion_purchase_allowed',
+        readonly=True
     )
-
-    @api.onchange('partner_id')
-    def onchange_partner_id_supplier_promotion_purchase_allowed(self):
-        self.supplier_promotion_allowed = (
-            self.partner_id.supplier_promotion_purchase_allowed
-        )