+++ HISTORY.rst
@@ -36,6 +36,7 @@ Unreleased
   * On lots change ref to product quantity
 * ALCYN-486: When creating a new sale order coming from the web service, make sure it is set with the appropriate default values (calling the existing onchange methods)
 * ALCYN-489: Change web service stock/product so it understand array how skus are sent in parameters
+* ALCYN-467: Use jobs to update promotion on purchase orders when the procurement is running
 
 **Build**